+++ packages/cip30/package.json
@@ -32,7 +32,6 @@
     "shx": "^0.3.3"
   },
   "dependencies": {
-    "@cardano-sdk/web-extension": " 0.2.0",
     "ts-custom-error": "^3.2.0",
     "ts-log": "^2.2.3",
     "webextension-polyfill": "^0.8.0"
