+++ packages/cip30/src/AuthenticatorApi/index.ts
@@ -1,5 +1,3 @@
 export * from './PersistentAuthenticator';
 export * from './PersistentAuthenticatorStorage';
-export * from './consumeRemoteAuthenticatorApi ';
-export * from './exposeAuthenticatorApi';
 export * from './types';
