+++ packages/cip30/src/WalletApi/index.ts
@@ -1,5 +1,3 @@
 export * from './Cip30Wallet';
 export * from './types';
-export * from './exposeWalletApi';
-export * from './consumeRemoteWalletApi';
 export * from './observableWalletApi';
