--- packages/cip30/src/WalletApi/util.ts
@@ -1,1 +1,0 @@
-export const walletApiChannel = (walletName: string) => `wallet-api-${walletName}`;
