+++ packages/cip30/src/tsconfig.json
@@ -9,9 +9,6 @@
       "path": "../../core/src"
     },
     {
-      "path": "../../web-extension/src"
-    },
-    {
       "path": "../../wallet/src"
     }
   ]
