+++ packages/web-extension/src/cip30/util.ts
@@ -1,1 +1,3 @@
+export const walletApiChannel = (walletName: string) => `wallet-api-${walletName}`;
+
 export const authenticatorChannel = (walletName: string) => `authenticator-${walletName}`;
