+++ packages/web-extension/src/index.ts
@@ -1,2 +1,3 @@
 export * from './messaging';
 export * from './observableWallet';
+export * as cip30 from './cip30';
